+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Add iosxe_interface_nve resource and data source
 - Add iosxe_evpn resource and data source
 - Add iosxe_evpn_instance resource and data source
+- Add iosxe_bgp resource and data source
 
 ## 0.1.0